+++ apps/backend/src/modules/activities/activity.repository.ts
@@ -155,6 +155,7 @@ class ActivityRepository implements Repository<ActivityEntity> {
 					.select("followingId")
 					.where({ followerId: userId }),
 			)
+			.orWhere(`${DatabaseTableName.ACTIVITIES}.userId`, userId)
 			.withGraphJoined(
 				`${RelationName.USER}.[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
 			)
@@ -248,6 +249,9 @@ class ActivityRepository implements Repository<ActivityEntity> {
 			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getLikesCountQuery())
 			.select(`${DatabaseTableName.ACTIVITIES}.*`, this.getCommentsCountQuery())
 			.returning("*")
+			.withGraphJoined(
+				`${RelationName.USER}.[${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}, ${RelationName.GROUPS}.${RelationName.PERMISSIONS}]`,
+			)
 			.castTo<(ActivityModel & ActivityCounts) | undefined>()
 			.execute();