+++ packages/shared/src/libs/enums/exception-message.enum.ts
@@ -14,6 +14,7 @@ const ExceptionMessage = {
 	NO_FILE_PRESENTED: "No file presented or format is incorrect.",
 	NO_PERMISSION: "No permission.",
 	NOTIFICATION_NOT_FOUND: "Notification not found.",
+	PROGRESS_NOT_FOUND: "Progress not found.",
 	SECTION_STATUS_NOT_FOUND: "Section status not found.",
 	TOKEN_EXPIRED: "Token expired.",
 	UNAUTHORIZED: "Unauthorized.",