+++ packages/shared/src/modules/user-notifications/enums/enums.ts
@@ -1,3 +1,4 @@
+export { NotificationFilter } from "./notification-filter.enum.js";
 export { NotificationValidationMessage } from "./notification-validation-message.enum.js";
 export { NotificationValidationRule } from "./notification-validation-rule.enum.js";
 export { UserNotificationsApiPath } from "./user-notifications-api-path.enum.js";