+++ packages/shared/src/modules/user-notifications/user-notifications.ts
@@ -1,3 +1,9 @@
-export { UserNotificationsApiPath } from "./enums/enums.js";
-export { type ReadNotificationsRequestDto } from "./types/types.js";
-export { readNotificationsRequest as readNotificationsRequestValidationSchema } from "./validation-schemas/validation-schemas.js";
+export { NotificationFilter, UserNotificationsApiPath } from "./enums/enums.js";
+export {
+	type NotificationFilterRequestDto,
+	type ReadNotificationsRequestDto,
+} from "./types/types.js";
+export {
+	readNotificationsRequest as readNotificationsRequestValidationSchema,
+	userNotificationQueryParameters as userNotificationQueryParametersValidationSchema,
+} from "./validation-schemas/validation-schemas.js";